move reference to active_session
